# Speed up parsing of annotations (XMLs)