@martvanrijthoven 2024-10-27
Hi Leander,

This is amazing, very nice speedup. Thank you so much!
I think there are no hurdles for moving to shapely2.0 and i think it should already work.